backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pandas>=1.5.0
openpyxl>=3.1.0
python-docx>=0.8.11
httpx>=0.24.0
aiofiles>=23.0.0
pydantic>=2.0.0
pytest>=8.0.0
